// socAPI/src/enums/UserRoleEnum.ts
/**
 * @description Enum representing user roles in the application. This enum is used to define the different roles that users can have, such as ADMIN, USER, and SUPER_ADMIN. Each role can be associated with specific permissions and access levels within the application, allowing for role-based access control and management of user privileges.
 */
export enum UserRoleEnum {
    ADMIN = 'ADMIN',
    USER = 'USER',
    SUPER_ADMIN = 'SUPER_ADMIN',
}